updated resume

+++ resume-builder-server/src/main/java/learn/resume/builder/domain/ResumeService.java
@@ -75,9 +75,16 @@ public class ResumeService {
         return result;
     }
 
-    //TODO: Update as Result
-    public boolean deleteByResumeId(int resumeId) {
-        return resumeRepo.deleteByResumeId(resumeId);
+
+    public Result deleteByResumeId(int resumeId) {
+        Result<Skill> result = new Result<>();
+        if (!resumeRepo.deleteByResumeId(resumeId)) {
+            result.addMessage("Resume Id was not found", ResultType.NOT_FOUND);
+        }
+        if(result.isSuccess()){
+            resumeRepo.deleteByResumeId(resumeId);
+        }
+        return result;
     }
 
     public Result updateResume(Resume resume) {
